+++ docs/release_notes.md
@@ -2,6 +2,10 @@
 ----------------
 
 
+- Analysis output file refactored.
+- Pull requests merged and issues closed:
+    - [PR #60](https://github.com/augustocattafesta/muGPD/pull/60)
+
 ### Version 0.0.2 (2026-03-20)
 
 - Added the package release notes;